requirements.txt
numpy
json
flask
pillow
tensorflow